working on flying camera

+++ src/main/java/com/grandopengame/engine/core/graphics/model/Model.java
@@ -1,12 +1,10 @@
 package com.grandopengame.engine.core.graphics.model;
 
 import lombok.Builder;
-import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
 import org.joml.Vector2f;
 import org.joml.Vector3f;
-import org.joml.Vector3i;
 
 import java.util.ArrayList;
 
@@ -18,5 +16,5 @@ public class Model {
     private ArrayList<Vector3f> vertices;
     private ArrayList<Vector3f> normals;
     private Vector2f[] uv;
-    private ArrayList<Short> indices;
+    private ArrayList<Integer> indices;
 }
